+++ RELEASE_NOTES.md
@@ -5,6 +5,18 @@ StandardButtons will use semver after 1.0.
 Until then, breaking changes can happen in any version, and deprecated features may be removed in any minor version bump.
 
 
+## 0.7
+
+This version adds standard actions and URLs. 
+
+### ✨ Features
+
+* `StandardAction` defines commonly used actions.
+* `StandardAction` can create a configured button.
+* `URL` has standard `.call(...)` and `.email(...)` builders.
+
+
+
 ## 0.6
 
 This version adds more button types and localized support.